+++ components/domain/map/Map/index.tsx
@@ -3,11 +3,10 @@
 import { Map as KakaoMap } from 'react-kakao-maps-sdk'
 
 import ErrorMessage from '@/components/common/ErrorMessage'
+import MapSkeleton from '@/components/common/Skeleton/MapSkeleton'
 import { MAP_CONFIG } from '@/constants/map'
 import { useKakaoLoader } from '@/hooks/useKakaoLoader'
 import { LatLng, MapRefType } from '@/types/map'
-
-import MapSkeleton from '../MapSkeleton'
 
 interface MapProps {
     ref?: MapRefType
